HTML and CSS Landing Page Practice
Description: This repository contains a landing page model that I created using HTML and CSS web development skills. The page is designed to be responsive.

The code files in this repository are the result of my practice and experimentation with HTML and CSS. I practice exercises to create this landing page.

Feel free to browse the code and use it as a reference for your own learning. If you have any feedback or suggestions for improvement, please create an issue or submit a pull request.



Thank you for visiting this repository and happy coding!